adds mean score to CVScores legend, removes ylim modification to make plots more comparable, updates from cv to cross_validation for readability
